+++ src/test/java/com/google/cloud/tools/appengine/cloudsdk/CloudSdkAppEngineStandardStagingTest.java
@@ -21,6 +21,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 import com.google.cloud.tools.appengine.api.AppEngineException;
+import com.google.cloud.tools.appengine.api.deploy.DefaultStageStandardConfiguration;
 import com.google.cloud.tools.appengine.api.deploy.StageStandardConfiguration;
 import com.google.cloud.tools.appengine.cloudsdk.process.ProcessHandlerException;
 import com.google.common.collect.ImmutableList;
@@ -47,7 +48,7 @@ public class CloudSdkAppEngineStandardStagingTest {
   private File destination;
   private File dockerfile;
   private CloudSdkAppEngineStandardStaging staging;
-  private StageStandardConfiguration.Builder builder;
+  private DefaultStageStandardConfiguration.Builder builder;
 
   @Before
   public void setUp() throws IOException {
@@ -58,7 +59,7 @@ public class CloudSdkAppEngineStandardStagingTest {
     staging = new CloudSdkAppEngineStandardStaging(appCfgRunner);
 
     builder =
-        new StageStandardConfiguration.Builder()
+        new DefaultStageStandardConfiguration.Builder()
             .setSourceDirectory(source)
             .setStagingDirectory(destination);
   }
